+++ src/routes/writing/_Article.svelte
@@ -1,14 +1,40 @@
 <script>
-	export let data = {};
+	import Webmentions from '../../components/Webmentions.svelte';
 
-	const {
-		description,
-		publishedAt,
-		readingTime,
-		slug,
-		title = 'Unknonw',
-	} = data;
+	export let description;
+	export let publishedAt;
+	export let readingTime;
+	export let slug;
+	export let title = 'Unknonw';
 </script>
+
+<style>
+	article :global(h1 + p) {
+		margin-left: 4em;
+		margin-bottom: 2em;
+	}
+
+	article :global(h2),
+	article :global(h3),
+	article :global(h4),
+	article :global(h5),
+	article :global(h6) {
+		text-transform: initial;
+	}
+
+	article :global(code) {
+		all: unset;
+	}
+
+	article :global(pre) {
+		max-width: 100ch;
+	}
+
+	/* break up long github issue links */
+	article :global(a) {
+		word-wrap: break-word;
+	}
+</style>
 
 <svelte:head>
 	<title>{title}</title>
@@ -36,4 +62,16 @@
 	<link
 		rel="webmention"
 		href="https://webmention.io/solverfox.dev/webmention" />
+	<link
+		rel="stylesheet"
+		href="https://highlightjs.org/static/demo/styles/a11y-dark.css" />
 </svelte:head>
+
+<article aria-details="webmentions">
+	<slot />
+</article>
+
+<hr />
+<Webmentions
+	id="webmentions"
+	target={`https://solverfox.dev/writing/${slug}`} />
