+++ src/routes/writing/testing-dispatch-event-pitfalls.svx
@@ -5,13 +5,8 @@ publishedAt: 2019-10-10
 description: Collection of common pitfalls when manually dispatch DOM events
 readingTime: 5
 slug: testing-dispatch-event-pitfalls
+layout: ./_Article.svelte
 ---
-
-```js exec
-import SEO from './_SEO.svelte';
-```
-
-<SEO data={_metadata} />
 
 # { \_metadata.title }
